+++ packages/press-ui/src/packages/mixins/uni-page-meta.d.ts
@@ -1,11 +1,11 @@
 export declare const uniSystemInfoMixin: {
-  data(): {
-    mixinRootFontSize: number;
-  };
   computed: {};
-  mounted(): void;
-  beforeDestroy(): void;
   methods: {
     onSetFontSize(): void;
   };
+  data(): {
+    mixinRootFontSize: number;
+  };
+  mounted(): void;
+  beforeDestroy(): void;
 };