+++ packages/press-ui/src/packages/mixins/visibility-change/inde.d.ts
@@ -1,7 +1,7 @@
 export declare const visibilityChangeMixin: (callback: Function) => {
-  mounted(this: any): void;
-  destroyed(this: any): void;
   methods: {
     _watchVisibleChange(): void;
   };
+  mounted(this: any): void;
+  destroyed(this: any): void;
 };