+++ packages/press-ui/src/packages/press-dialog/computed.js
@@ -73,4 +73,8 @@ export const dialogProps = {
     type: String,
     default: '',
   },
+  customClass: {
+    type: String,
+    default: '',
+  },
 };