+++ packages/press-ui/src/packages/press-dropdown-menu/computed.js
@@ -3,7 +3,7 @@ function displayTitle(item) {
     return item.title;
   }
 
-  const match = item.options.filter((option) => option.value === item.value);
+  const match = item.options.filter(option => option.value === item.value);
   const displayTitle = match.length ? match[0].text : '';
   return displayTitle;
 }